\name{data}
\alias{data}

\docType{data}

\title{Example data}

\description{
CSES data (modules I-III). The response variable is a matrix of party vote-shares across democracies. The covariate is the average number of electoral districts in a country.
}

\usage{data(data)}

\format{A list of vote-shares and the list of average electoral districts.}

\source{\url{www.cses.org}}

\keyword{datasets}
