\name{ocomposition-package}
\alias{ocomposition-package}
\alias{ocomposition}
\docType{package}
\title{ordered compositional data model}
\description{A Gibbs sampler for a Bayesian model of ordered compositional data (e.g., data on party systems).}
\details{
\tabular{ll}{
Package: \tab ocomposition\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-08-27\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
}
\author{
Arturas Rozenas, Duke University <ar71@duke.edu>
}
\references{Rozenas, Arturas (2011) "A Statistical Model for Party Systems Analysis", Political Analysis, in print.
}
\keyword{ package }
\seealso{}
\examples{}
