\name{summary.composition}
\alias{summary.composition}

\title{Summary function}

\description{
Summarizes the basic properties of the MCMC output: means, standard deviations, MCMC errors, and Heidel convergence diagnostics.
}

\usage{
\method{summary}{composition}(object, individual = NULL, \dots)
}

\arguments{
\item{object}{MCMC output list member of class \code{composition}, generated by call to \code{fitcomp}.}
\item{individual}{Integer from 1 to the number of components in data. Outputs summary for that component; if \code{NULL}, then outputs summaries of the hyperparameters.}
\item{\dots}{...}
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line